+++ runtime_src/calibration_profiles.js
@@ -1520,6 +1520,7 @@
 
   window.SwitchVisionCalibrationProfiles = {
     load,
+    deleteSelected,
     deleteAllUnused,
   };
 })();